Remove Promise as return from handler functions
In order to make the handler functions compatible with how AWS expects
them to work, handlers should either (1) invoke the provided callback
function with the result or (2) return a `Promise` containing the result.

Previously the handler function signatures prevented us from doing approach
nr 2, because the returned `Js.Promise(unit)` could not contain an object
with the result.

Therefore going for approch nr 1 where the provided callback should be
invoked with the result, and no return value out of the handler functions.