Remove callback argument from handler functions
These changes removes the callback argument from all the handler function
types, and changing the returned promises to contain the lambda result.

Main motivation of doing this is to make the handler function compatible
with how the AWS Lambda machinery expects handlers to work;

- Either invoke the provided callback function with the result
- Or return a `Promise` with the result

It's one or the other, not both as the previous handler type declarations
gave the impression of.

Refs https://github.com/ahrefs/bs-aws-lambda/issues/16